requirements.txt
faiss_cpu==1.12.0
fastapi==0.121.1
faster_whisper==1.2.0
librosa==0.11.0
numpy==2.3.4
Pillow==12.0.0
pydantic==2.12.4
PyMuPDF==1.26.4
pymupdf4llm==0.0.27
pyttsx3==2.99
PyYAML==6.0.3
rapidfuzz==3.14.3
requests==2.32.5
safetensors==0.6.2
sentence_transformers==5.1.1
sentencepiece==0.2.1
soundfile==0.13.1
spacy==3.7.5
streamlit==1.50.0
tiktoken==0.11.0
tqdm==4.67.1
transformers==4.57.0
uvicorn==0.38.0
